files/requirements.txt
flask
mysql-connector-python
requests
pytest
ansible
ansible-lint
